refactor: crete DependencyGraph wrapper
